requirements.txt
# FastAPI Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Validation
pydantic==2.10.3
pydantic[email]==2.10.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
